Add files via upload
Summary: The research looked to bridge the gap between access to the sports industry and resources to assist Latina women interested in sports careers and college sports participation. The studies included surveys and interviews focused on career goals, upbringing, cultural competency, and finance.

Abstract: This study, conducted in collaboration with the ELLA Sports Foundation, examined the
underrepresentation of Latina girls in sports. A total of 319 participants were recruited through
various social media platforms and took part in the survey. The survey included Likert-scale
Style items are based on three key constructs: collectivism, economics, and societal beliefs. It was
hypothesized that each construct would contribute to the need for more participation of Latina
girls in sports. The analysis revealed no significant differences between the responses of
Latino/Hispanic participants and participants of other ethnic backgrounds. The results suggested
the belief that sports participation is costly and time-consuming and encouraged increased
support for Latina girls in sports. The only concern that lowered participant support was the
belief that girls should only travel for a short time for sports. With safety being a priority, the
findings recommend implementing stringent background checks for chaperones and suggest that
enhancing funding and support could boost the participation rates of Latina girls in sports.
Keywords: Latina athletes, sports participation, collectivism, gender disparity, athletic
diversity, motivation, cultural influence, societal influence, economic factors